// types.d.ts
export type WebhookEvent = 'POST_PUBLISHED' | 'POST_UPDATED' | 'POST_DELETED' | 'STATIC_PAGE_PUBLISHED' | 'STATIC_PAGE_UPDATED' | 'STATIC_PAGE_DELETED';